PdfToPodcast/api/pdf-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.22
pypdf>=4.0.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.3.0
pdf2image==1.16.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-magic==0.4.27
langdetect==1.0.9
